code cleaning & reorg - final deprecation of public services, yahi - cache agent is more a prototype than a draft - move settings out of application in code root - make tests green again, deprecate old vo_service tests
